% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{find_package_name}
\alias{find_package_name}
\title{find_package_name}
\usage{
find_package_name()
}
\value{
Name of the package that user is in
}
\description{
Finds the name of the package that the user is working on.
}
