% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{get_varName}
\alias{get_varName}
\title{Find variable names in the output of R CMD check}
\usage{
get_varName(string)
}
\arguments{
\item{string}{The string we're looking for}
}
\value{
A list of matches
}
\description{
Find variable names in the output of R CMD check
}
