% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{append_file}
\alias{append_file}
\title{Read from and wrtie to zzz_global_variables.R}
\usage{
append_file(new_vars, file_path)
}
\arguments{
\item{new_vars}{A vector with names of new variables to add}

\item{file_path}{Path to zzz_global_variables.R}
}
\description{
Read from and wrtie to zzz_global_variables.R
}
